+++ src/main/java/me/flotsam/frettler/Main.java
@@ -25,6 +25,7 @@ import me.flotsam.frettler.command.CompletionsCommand;
 import me.flotsam.frettler.command.FrettlerCommand;
 import me.flotsam.frettler.command.GuitarCommand;
 import me.flotsam.frettler.command.MandolinCommand;
+import me.flotsam.frettler.command.PatternsCommand;
 import me.flotsam.frettler.command.UkeleleCommand;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
@@ -32,7 +33,7 @@ import picocli.CommandLine.HelpCommand;
 
 @Command(name = "frettler",
     description = "Generates scales/arpeggios, chords and arpeggios for fretted instruments",
-    subcommands = {HelpCommand.class, CompletionsCommand.class, ChordCommand.class, GuitarCommand.class,
+    subcommands = {HelpCommand.class, PatternsCommand.class, CompletionsCommand.class, ChordCommand.class, GuitarCommand.class,
         BassGuitarCommand.class, BanjoCommand.class, MandolinCommand.class, UkeleleCommand.class})
 public class Main implements Callable<Integer> {